refactor(Entity Types): Moved all types from entity files to /types dir
