+++ packages/server/src/entities/PermissionLevel.ts
@@ -1,22 +1,9 @@
-import { ObjectType, Field, registerEnumType } from "type-graphql";
+import { ObjectType, Field } from "type-graphql";
 import { Entity, Column } from "typeorm";
 
 import { tableNames } from "../constants/tableNames";
+import { PermissionLevelNameEnum } from "../types/permissionLevel.types";
 import { DefaultColumns } from "./common/DefaultColumns";
-
-export enum PermissionLevelIdEnum {
-  Admin = 1,
-  User = 2,
-  Guest = 3,
-}
-enum PermissionLevelNameEnum {
-  Admin = "ADMIN",
-  User = "USER",
-  Guest = "GUEST",
-}
-registerEnumType(PermissionLevelIdEnum, {
-  name: "PermissionLevelEnum",
-});
 
 @ObjectType()
 @Entity(tableNames.permissionLevel)
