+++ packages/server/src/types/user.types.ts
@@ -1,0 +1,4 @@
+export enum UserAccountStatusEnum {
+  NotConfirmed = "NOT_CONFIRMED",
+  Confirmed = "CONFIRMED",
+}
